api/requirements.txt
flask
Flask
mysql-connector
redis
pykafka
